+++ HISTORY.rst
@@ -1,3 +1,12 @@
+3.4.0 (Sep 27, 2016)
+~~~~~~~~~~~~~~~~~~~~
+
+- Added asyncio support. (Python 3.5+)
+- Added a *methods* object to the jsonrpcserver module (so you can import
+  jsonrpcserver.methods, rather than instantiating your own).
+- Added methods.dispatch().
+- *A ton of refactoring.*
+
 3.3.4 (Sep 22, 2016)
 ~~~~~~~~~~~~~~~~~~~~